[LiveChat] migrated the models from games to chat and added manytomany field for users
