+++ src/chigame/games/models.py
@@ -691,35 +691,6 @@ class Review(models.Model):
         super().save(*args, **kwargs)
 
 
-class LiveChat(models.Model):
-    """
-    Represents a new live chat between users.
-    """
-
-    # used to identify which channel the chat is on
-    channel = models.TextField(unique=True, null=False)
-
-
-class LiveChatMessage(models.Model):
-    """
-    Represents a new message in the live chat.
-    """
-
-    live_chat_id = models.ForeignKey(LiveChat, on_delete=models.CASCADE)
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    sent_at = models.DateTimeField(auto_now_add=True)
-    message_content = models.TextField(null=False)
-
-
-class LiveChatUser(models.Model):
-    """
-    Represents a user mapped to a new live chat.
-    """
-
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    live_chat_id = models.ForeignKey(LiveChat, on_delete=models.CASCADE)
-
-
 class GameList(models.Model):
     """
     A collection of games defined by users.
